+++ protected/humhub/components/access/ControllerAccess.php
@@ -494,7 +494,7 @@ class ControllerAccess extends BaseObject
     public function validateDisabledUser()
     {
         return $this->isGuest()
-            || ($this->user->status !== User::STATUS_DISABLED
+            || ($this->user->status !== User::STATUS_DEACTIVATED
                 && $this->user->status !== User::STATUS_SOFT_DELETED);
     }